rename agent abc file to abc
