+++ ubikagent/agent/__init__.py
@@ -1,4 +1,4 @@
-from .agent import Agent
+from .abc import Agent
 from .random import RandomAgent, UnityRandomAgent
 from .dqn import DQNAgent
 from .ddpg import DDPGAgent
